Move model objects to their own source files
