+++ Sources/PushNotifications/Models/PublishResponse.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct PublishResponse: Decodable {
+    let id: String
+
+    enum CodingKeys: String, CodingKey {
+        case id = "publishId"
+    }
+}
