+++ Sources/PushNotifications/Models/SDKVersion.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+/// The version number of the SDK.
+struct SDKVersion {
+
+    /// The current version number `String`.
+    ///
+    /// The version number adheres to Semantic Versioning: http://semver.org/spec/v2.0.0.html .
+    static let current = "1.0.3"
+}
